feat: #1415; endpoint; scores-lib; player-links;
