+++ src/lib/components/page/home/top-goalscorers/TopGoalScorers-Loader.svelte
