+++ src/lib/components/page/home/top-goalscorers/loaders/_Placeholder_Table_Row.svelte
