+++ src/lib/components/page/home/top-goalscorers/loaders/desktop/_Placeholder_Row.svelte
