+++ src/lib/components/page/home/top-goalscorers/loaders/mobile/_Placeholder_Row.svelte
