+++ src/lib/components/page/home/top-goalscorers/loaders/tablet/_Placeholder_Row.svelte
